Refactor roster view into modular components
