+++ starbase/ai-roomchat/components/roster/RosterView/styles.js
@@ -1,243 +1,3 @@
-'use client'
-
-import Link from 'next/link'
-
-import LogoutButton from '../LogoutButton'
-
-const COMMUNITY_URL = 'https://gall.dcinside.com/mini/board/lists/?id=gionkirr'
-
-function formatDate(value) {
-  if (!value) return '—'
-
-  const date = new Date(value)
-  if (Number.isNaN(date.getTime())) return '—'
-
-  return date.toLocaleDateString('ko-KR', {
-    month: 'short',
-    day: 'numeric',
-  })
-}
-
-function CreateHeroCard() {
-  return (
-    <Link href="/create" passHref>
-      <a style={styles.createCardLink}>
-        <div style={styles.createCardCover}>
-          <div style={styles.createCardOverlay}>
-            <div style={styles.createCardIconWrapper}>
-              <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
-                <path d="M11 5h2v6h6v2h-6v6h-2v-6H5v-2h6V5Z" fill="currentColor" />
-              </svg>
-            </div>
-            <div style={styles.createCardTextGroup}>
-              <span style={styles.createCardTitle}>영웅 생성</span>
-              <span style={styles.createCardSubtitle}>지금 영웅을 소환해보세요</span>
-            </div>
-          </div>
-        </div>
-      </a>
-    </Link>
-  )
-}
-
-function HeroCard({ hero, onDelete }) {
-  return (
-    <Link href={`/character/${hero.id}`} passHref>
-      <a style={styles.heroCardLink}>
-        <div style={styles.heroCardCover}>
-          <button
-            type="button"
-            onClick={(event) => {
-              event.preventDefault()
-              event.stopPropagation()
-              onDelete(hero)
-            }}
-            onMouseDown={(event) => event.stopPropagation()}
-            style={styles.heroDeleteButton}
-            aria-label={`${hero.name} 삭제`}
-          >
-            <svg width="18" height="18" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
-              <path d="M9 3h6l1 2h4v2H4V5h4l1-2Zm1 6h2v8h-2V9Zm4 0h2v8h-2V9Z" fill="currentColor" />
-            </svg>
-          </button>
-          {hero.image_url ? (
-            // eslint-disable-next-line @next/next/no-img-element
-            <img src={hero.image_url} alt={hero.name} style={styles.heroImage} />
-          ) : (
-            <div style={styles.heroImageFallback}>{hero.name?.slice(0, 2) ?? '??'}</div>
-          )}
-          <div style={styles.heroCardFooter}>
-            <div style={styles.heroName}>{hero.name}</div>
-            <span style={styles.heroCreatedAt}>{formatDate(hero.created_at)}</span>
-          </div>
-        </div>
-      </a>
-    </Link>
-  )
-}
-
-function HeroList({ loading, error, heroes, onRequestDelete, onResetError }) {
-  const createCard = <CreateHeroCard />
-
-  if (loading) {
-    return <div style={styles.loadingState}>로스터를 불러오는 중입니다…</div>
-  }
-
-  if (error) {
-    return (
-      <>
-        <div style={styles.errorBox} role="alert">
-          {error}
-          <button type="button" onClick={onResetError} style={styles.errorResetButton}>
-            다시 시도
-          </button>
-        </div>
-        {createCard}
-      </>
-    )
-  }
-
-  if (!heroes.length) {
-    return (
-      <>
-        <div style={styles.emptyState}>
-          아직 등록된 영웅이 없습니다.
-          <br />
-          아래의 <span style={{ color: '#38bdf8' }}>영웅 생성</span> 카드를 눌러 첫 캐릭터를 만들어보세요.
-        </div>
-        {createCard}
-      </>
-    )
-  }
-
-  return (
-    <>
-      {heroes.map((hero) => (
-        <HeroCard key={hero.id} hero={hero} onDelete={onRequestDelete} />
-      ))}
-      {createCard}
-    </>
-  )
-}
-
-function DeleteHeroModal({ hero, open, deleting, onCancel, onConfirm }) {
-  if (!open || !hero) return null
-
-  return (
-    <div style={styles.modalBackdrop} role="dialog" aria-modal="true">
-      <div style={styles.modalBody}>
-        <div style={styles.modalIcon}>
-          <svg width="32" height="32" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
-            <path d="M9 3h6l1 2h4v2H4V5h4l1-2Zm1 6h2v8h-2V9Zm4 0h2v8h-2V9Z" fill="currentColor" />
-          </svg>
-        </div>
-        <div style={styles.modalTextGroup}>
-          <h2 style={styles.modalTitle}>영원히 삭제할까요?</h2>
-          <p style={styles.modalSubtitle}>
-            삭제하면 3분간 새로 생성할 수 없습니다.
-            <br />
-            <strong>{hero.name}</strong> 영웅을 삭제하시겠습니까?
-          </p>
-        </div>
-        <div style={styles.modalActions}>
-          <button type="button" onClick={onCancel} style={styles.modalCancelButton} disabled={deleting}>
-            취소
-          </button>
-          <button
-            type="button"
-            onClick={onConfirm}
-            style={{
-              ...styles.modalDeleteButton,
-              background: deleting
-                ? 'rgba(248,113,113,0.5)'
-                : 'linear-gradient(135deg, #fb7185 0%, #ef4444 100%)',
-              cursor: deleting ? 'not-allowed' : 'pointer',
-              opacity: deleting ? 0.7 : 1,
-            }}
-            disabled={deleting}
-          >
-            삭제
-          </button>
-        </div>
-      </div>
-    </div>
-  )
-}
-
-export default function RosterView({
-  loading,
-  error,
-  heroes,
-  displayName,
-  avatarUrl,
-  deleteTarget,
-  deleting,
-  onRequestDelete,
-  onCancelDelete,
-  onConfirmDelete,
-  onLogoutComplete,
-  onResetError,
-}) {
-  return (
-    <div style={styles.page}>
-      <div style={styles.container}>
-        <header style={styles.header}>
-          <div style={styles.headerTopRow}>
-            <div style={styles.titleGroup}>
-              <div style={styles.badge}>Tale of Heroes</div>
-              <h1 style={styles.title}>나의 영웅 도감</h1>
-              <p style={styles.subtitle}>영웅들을 관리하고 로비에서 곧바로 자랑해 보세요.</p>
-            </div>
-            <LogoutButton avatarUrl={avatarUrl} displayName={displayName} onAfter={onLogoutComplete} />
-          </div>
-
-          <a href={COMMUNITY_URL} target="_blank" rel="noreferrer" style={styles.communityCard}>
-            <span style={styles.communityLabel}>공식 커뮤니티 오픈!</span>
-            <p style={styles.communityDescription}>
-              캐릭터 자랑, 설정 공유, 팬아트까지 모두 환영합니다. 지금 바로 커뮤니티에서 첫 인사를 남겨보세요.
-            </p>
-          </a>
-
-          <div style={styles.heroCallout}>
-            <div style={styles.heroCalloutOverlay} />
-            <div style={styles.heroCalloutContent}>
-              <span style={styles.heroCalloutEyebrow}>영웅을 생성하고</span>
-              <h2 style={styles.heroCalloutTitle}>전설을 시작하세요</h2>
-              <p style={styles.heroCalloutSubtitle}>
-                새로운 영웅을 만들고 스토리를 기록하면 플레이어들이 당신의 세계를 함께 즐기게 됩니다.
-              </p>
-            </div>
-          </div>
-        </header>
-
-        <section style={styles.heroSection}>
-          <div style={styles.heroSectionHeader}>
-            <h2 style={styles.heroSectionTitle}>내 영웅 목록</h2>
-            <span style={styles.heroSectionCount}>{heroes.length}명</span>
-          </div>
-          <div style={styles.heroListContainer}>
-            <HeroList
-              loading={loading}
-              error={error}
-              heroes={heroes}
-              onRequestDelete={onRequestDelete}
-              onResetError={onResetError}
-            />
-          </div>
-        </section>
-      </div>
-
-      <DeleteHeroModal
-        hero={deleteTarget}
-        open={Boolean(deleteTarget)}
-        deleting={deleting}
-        onCancel={onCancelDelete}
-        onConfirm={onConfirmDelete}
-      />
-    </div>
-  )
-}
-
 const styles = {
   page: {
     minHeight: '100vh',
@@ -639,3 +399,5 @@ const styles = {
     boxShadow: '0 18px 48px -28px rgba(248,113,113,0.95)',
   },
 }
+
+export default styles
